app/server/app_backend/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
annotated-types==0.7.0
attrs==25.3.0
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
datasets==4.0.0
-e git+https://github.com/huggingface/diffusers@67ffa7031e5a4bf0991b692a424e36ca59e64ec9#egg=diffusers
dill==0.3.8
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
grpcio==1.74.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
lpips==0.1.4
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
peft==0.17.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
scipy==1.16.2
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
stringzilla==4.2.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.20
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
xformers==0.0.32.post2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
